Fixed logic to delete metadata restriction on a field only if it exists. Updated api call for downloading a sequencing file from the sequencing runs details page

+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/services/UISampleService.java
@@ -403,14 +403,17 @@ public class UISampleService {
 		Long fieldUsageCount = metadataEntryRepository.getMetadataEntriesCountBySamplesAndField(metadataTemplateField,
 				sampleList);
 		metadataEntryRepository.deleteById(metadataEntryId);
-		MetadataRestriction restrictionToDelete = metadataTemplateService.getMetadataRestrictionForFieldAndProject(
-				project, metadataTemplateField);
+
 		/*
 		 Only delete the restriction on the field if there is only one place
 		 where the field is in use within the project
 		 */
 		if (fieldUsageCount == 1) {
-			metadataRestrictionRepository.delete(restrictionToDelete);
+			MetadataRestriction restrictionToDelete = metadataTemplateService.getMetadataRestrictionForFieldAndProject(
+					project, metadataTemplateField);
+			if(restrictionToDelete != null) {
+				metadataRestrictionRepository.delete(restrictionToDelete);
+			}
 		}
 
 		return messageSource.getMessage("server.sample.metadata.remove.success",
